requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
apsw==3.45.1.0
argcomplete==3.5.3
asgiref==3.8.1
attrs==25.1.0
Automat==24.8.1
babel==2.17.0
Beaker==1.12.1
beautifulsoup4==4.12.3
blinker==1.7.0
blivet==3.10.0
blivet-gui==2.6.0
boto3==1.36.22
botocore==1.36.22
Brlapi==0.8.5
Brotli==1.1.0
cairocffi==1.6.1
CairoSVG==2.7.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
constantly==23.10.4
cryptography==41.0.7
css-parser==1.0.7
cssselect==1.2.0
cssselect2==0.7.0
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
Django==5.0.4
dnf==4.22.0
dnspython==2.6.1
docx==0.2.4
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
fedora-third-party==0.10
feedparser==6.0.11
file-magic==0.4.0
filelock==3.16.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fros==1.1
fsspec==2024.9.0
FVS==0.3.4
googletrans==3.0.0
greenlet==3.0.3
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
html2text==2024.2.25
html5-parser==0.4.12
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humanize==3.13.1
hyperframe==5.2.0
hyperlink==21.0.0
icoextract==0.1.5
idna==2.10
ifaddr==0.1.7
incremental==24.7.2
initial-setup==0.3.100
instaloader==4.11
iso639==0.1.4
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.5
jmespath==1.0.1
langtable==0.0.68
libcomps==0.1.20
libdnf==0.73.4
libretranslatepy==2.1.1
llvmlite==0.43.0
louis==3.28.0
lxml==5.1.0
Mako==1.2.3
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mechanize==0.4.8
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.0.6
mutagen==1.47.0
netifaces==0.11.0
networkx==3.3
nftables==0.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
odfpy==1.4.1
olefile==0.47
openai-whisper @ git+https://github.com/openai/whisper.git@279133e3107392276dc509148da1f41bfb532c7e
orjson==3.10.14
packaging==23.2
parsel==1.10.0
Paste==3.7.1
pathvalidate==3.2.0
patool==1.13
pefile==2023.2.7
pexpect==4.9.0
pid==2.2.3
pillow==10.3.0
platformdirs==4.3.6
ply==3.11
plyer==2.1.0
productmd==1.43
progressbar2==3.53.2
Protego==0.4.0
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pwquality==1.4.5
pyaes==1.6.1
pyasn1==0.6.0
pyasn1_modules==0.4.1
pycairo==1.25.1
pychm==0.8.6
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.20.0
pycups==2.0.4
pycurl==7.45.2
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyDispatcher==2.0.7
pyenchant==3.2.2
pygit2==1.14.0
Pygments==2.17.2
PyGObject==3.48.2
pykickstart==3.52
pyOpenSSL==23.2.0
pyparted==3.13.0
PyPDF2==3.0.1
PyQt6==6.8.1.dev2502011625
PyQt6-WebEngine==6.8.0
PyQt6_sip==13.9.1
PySocks==1.7.1
pystray==0.17.3
pytesseract==0.3.10
python-augeas==1.1.0
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-gettext==4.0
python-manatools==0.0.4
python-meh==0.51
python-multipart==0.0.20
python-pam==2.0.2
python-utils==3.7.0
python-xlib==0.33
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
queuelib==1.7.0
regex==2024.9.11
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
rfc3986==1.5.0
rich==13.9.4
rich-toolkit==0.13.2
rpm==4.19.1.1
rpmautospec==0.7.3
rpmautospec-core==0.1.5
rsa==4.9
s3transfer==0.11.2
Scrapy==2.12.0
selinux @ file:///builddir/build/BUILD/libselinux-3.7/src
sepolicy @ file:///builddir/build/BUILD/selinux-3.7/python/sepolicy
service-identity==24.2.0
setools==4.5.1
setuptools==69.0.3
sgmllib3k==1.0.0
shellingham==1.5.4
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.1
sos==4.7.2
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.5.0
starlette==0.46.0
sympy==1.13.2
systemd-python==235
Telethon==1.34.0
Tempita==0.5.2
tiktoken==0.7.0
tinycss2==1.2.1
tldextract==5.1.3
torch==2.4.1
tqdm==4.66.5
translate==3.6.1
triton==3.0.0
Twisted==24.11.0
typer==0.15.2
typing_extensions==4.12.2
ufw==0.35
ujson==5.10.0
urllib3==1.26.20
uvicorn==0.23.2
uvloop==0.21.0
virtualenv==20.29.2
vkbasalt-cli==3.1.1.post2
w3lib==2.3.1
watchfiles==1.0.4
webencodings==0.5.1
websockets==13.0.1
Werkzeug==3.0.2
xkbregistry==0.3
xxhash==3.5.0
yt-dlp==2024.8.6
zeroconf==0.118.0
zope.interface==7.2
